Monstrous patch from Ralf Corsepius <corsepiu@faw.uni-ulm.de>.  I have made no attempt to divide the comments up and place them with just the appropriate files.  Here is an excerpt from Ralf's email:
Changes including comments on changes I made after cycling through
all the targets:

  * Added ranlib support. Now all targets use "ranlib" instead of "ar -s"
    to build an index for a library. If ranlib isn't detected during
    configuration, check if ar -s is working and try "ar -s" instead of

  * Removed $(XXX_FOR_TARGET) from make/target.cfg.in, use $(XXX) instead now.

  * gcc-target-default.cfg: LINK_XXXX-defines reworked to solve the -l
    problem under posix (cf gcc-target-default.cfg)

  * rtems-glom replaced by Makefile-rules inside of the wrapup/Makefile.in
    that has been using rtems-glom until now.

  * Removed CCC and friends in gcc-target-default.cfg, as they have been
    breaking CXX support.

  * Removed CONFIG.$(TARGET_ARCH).CC lines from several custom/*.cfg
    files, because this is now set in custom/default.cfg.

  * Added aclocal/ar-s.m4, check whether "ar -s" is working

  * Added aclocal/cygwin.m4 and aclocal/exeext.m4.

  * Reworked aclocal/canonicalize-tools.m4: Added ar -s check; fixes for
    problems when  XXX_FOR_TARGET is given via environment variables (didn't
    work for gcc until now), adding cygwin check, improved autoconf-cache
    handling.

  * Removed -l from make rule dependencies. LINK_LIBS is now allowed to
    contain -L and -l. LINK_OBJS and LINK_FILES must not contain -L or -l.
    gcc28 make-exe rules now link using $(LINK_OBJS) $(LINK_LIBS) => Almost
    all custom/*.cfg are modified. This is very likely to break something
    because of typos or having missed to edit a file.

  Open problems, known bugs, things I didn't do:

  * custom/p4000.cfg seems to be out of date and requires to be reviewed.

    (JRS NOTE: It is subordinate p4650 and p4600 -- both of which build ok
               after minor changes.)

  * custom/psim.cfg needs to be reviewed, I added some changes to it, I am
    insecure about.

    (JRS NOTE: psim had a minor problem endif/endef swapped but runs fine.)

  * rtems-glom.in can now be removed.

  * gcc*.cfg files "make depend" rules don't honor language specific flags
    (e.g CXXFLAGS is ignored for *.cc) - Nothing to worry about now, but may
    cause problems for hosts/targets not using gcc or rtems-add-ons that use
    external packages.

  * AFAIS, the no_bsp BSP can't be build anymore, i.e. configure refused
    to configure for it whatever I tried.

  * The toplevel and toplevel+1 README files are quite out-dated

  * cygwin.m4 isn't of much use for rtems. In most cases (cf.
    aclocal/*.m4) it is worked around by directly using $host_os. I think
    I'll remove it soon after the next snapshot

  * Before release the cygwin patch needs to be tested under cygwin. I may
    have broken/missed something (esp. the sed-pattern to convert \\ into /
    may be broken).

  * You should try to build/run the posix-BSP under solaris - I don't
    expect problems, but I am not 100% sure, esp. with regard to ranlib/ar -s.

  * You should consider to convert all make/compilers/*.cfg files into
    make/compilers/*.cfg.in files and let autoconf generate the *.cfg. This
    may help getting rid of some if/then/else statements and help
    hard-coding some defines into those files in future and shouldn't
    disturb now.

  * Not having installed libc.a/libm.a on a host may still break building
    rtems, esp. when using -disable-gcc28 as the gcc27-configuration scheme
    directly accesses libc.a and libm.a. The problem should not appear when
    using gcc28 because it references libc/libm only through -lc and -lm
    which may be static or dynamic (I didn't test this).

  * shgen is not yet included (I didn't yet have enough time to integrate it).

  * I know about a few more configure-probs (esp. cross-checking
    --enable-* flags).
     + warn/refuse to configure when --enable-libcdir and
       --enable-gcc28 are given.
     + force --enable-libcdir when --disable-gcc28 is given

  * Replaced KSHELL with @KSH@ in some shell scripts generated by configure.in.

  * Added a dependency to aclocal/*.m4 in the toplevel Makefile => configure
    and aclocal.m4 will now be rebuild when any aclocal/*.m4 file is changed

  * Some changes to aclocal/gcc-pipe.m4 and aclocal/gcc-specs.m4

  * Replaced i[[3456]]86-unknown-freebsd2.[[12]] with i[[3456]]86-*freebsd2.*
    in configure.in, as I suppose there might exist a variety of valid vendors
    (2nd field of the name-tripple)

  * Disabled override MAKEFLAGS in toplevel Makefile.in - Potential
    side-effects are not really clear to me.

  * In mvme162.cfg, $(LINK_LIBS) is missing in the CC line in gcc28's make-exe
    rule (yet another one I missed to edit). Just append $(LINK_LIBS) to
    the "CC" line, like I hopefully did to ALL other custom/*.cfg files.

  * the problem with mvme162lx.cfg is a follow-up problem of the
    mvme162.cfg-bug.

  * mvme162/console and idp/console had variables named Buffer which
    conflicted with similarly named variables in some tests.

+++ aclocal.m4
@@ -151,12 +151,14 @@ dnl
 AC_DEFUN(RTEMS_GCC_SPECS,
 [AC_REQUIRE([RTEMS_PROG_CC])
 AC_CACHE_CHECK(whether $CC_FOR_TARGET accepts -specs,rtems_cv_gcc_specs,
-[touch confspec
-echo 'void f(){}' >conftest.c
-if test -z "`${CC_FOR_TARGET} -specs confspec -c conftest.c 2>&1`";then
-  rtems_cv_gcc_specs=yes
-else
-  rtems_cv_gcc_specs=no
+[
+rtems_cv_gcc_specs=no
+if test "$rtems_cv_prog_gcc" = "yes"; then
+  touch confspec
+  echo 'void f(){}' >conftest.c
+  if test -z "`${CC_FOR_TARGET} -specs confspec -c conftest.c 2>&1`";then
+    rtems_cv_gcc_specs=yes
+  fi
 fi
 rm -f confspec conftest*
 ])])
@@ -174,20 +176,20 @@ AC_DEFUN(RTEMS_GCC_PIPE,
 AC_REQUIRE([AC_CANONICAL_HOST])
 AC_CACHE_CHECK(whether $CC_FOR_TARGET accepts --pipe,rtems_cv_gcc_pipe,
 [
+rtems_cv_gcc_pipe=no
+if test "$rtems_cv_prog_gcc" = "yes"; then
 case "$host_os" in
   cygwin32*)
-    rtems_cv_gcc_pipe=no 
     ;;
   *)
     echo 'void f(){}' >conftest.c
     if test -z "`${CC_FOR_TARGET} --pipe -c conftest.c 2>&1`";then
       rtems_cv_gcc_pipe=yes
-    else
-      rtems_cv_gcc_pipe=no
     fi
     rm -f conftest*
     ;;
 esac
+fi
 ])
 ])
 
@@ -261,40 +263,174 @@ dnl $Id$
 dnl
 dnl Set target tools
 dnl
+dnl 98/06/23 Ralf Corsepius	(corsepiu@faw.uni-ulm.de)
+dnl		fixing cache/environment variable handling
+dnl		adding checks for cygwin/egcs '\\'-bug
+dnl		adding checks for ranlib/ar -s problem 
+dnl
 dnl 98/02/12 Ralf Corsepius	(corsepiu@faw.uni-ulm.de)
 dnl
 
+AC_DEFUN(RTEMS_GCC_PRINT,
+[ case $host_os in
+  *cygwin32*)
+    dnl FIXME: Hack for cygwin/egcs reporting mixed '\\' and '/'
+    dnl        Should be removed once cygwin/egcs reports '/' only
+    $1=`$CC_FOR_TARGET --print-prog-name=$2 | sed -e "s%\\\\%/%g" `
+    ;;
+  *)
+    $1=`$CC_FOR_TARGET --print-prog-name=$2`
+    ;;
+  esac
+])
+
+AC_DEFUN(RTEMS_PATH_TOOL,
+[
+AC_MSG_CHECKING([target's $2])
+AC_CACHE_VAL(ac_cv_path_$1,:)
+AC_MSG_RESULT([$ac_cv_path_$1])
+
+if test -n "$ac_cv_path_$1"; then
+  dnl retrieve the value from the cache
+  $1=$ac_cv_path_$1
+else
+  dnl the cache was not set
+  if test -z "[$]$1" ; then
+    if test "$rtems_cv_prog_gcc" = "yes"; then
+      # We are using gcc, ask it about its tool
+      # NOTE: Necessary if gcc was configured to use the target's 
+      # native tools or uses prefixes for gnutools (e.g. gas instead of as)
+      RTEMS_GCC_PRINT($1,$2)
+    fi
+  else
+    # The user set an environment variable.
+    # Check whether it is an absolute path, otherwise AC_PATH_PROG 
+    # will override the environment variable, which isn't what the user
+    # intends
+    AC_MSG_CHECKING([whether environment variable $1 is an absolute path])
+    case "[$]$1" in
+    /*) # valid
+      AC_MSG_RESULT("yes")
+    ;;
+    *)  # invalid for AC_PATH_PROG
+      AC_MSG_RESULT("no")
+      AC_MSG_ERROR([***]
+        [Environment variable $1 should ether]
+	[be unset (preferred) or contain an absolute path])
+    ;;
+    esac
+  fi
+
+  AC_PATH_PROG($1,"$program_prefix"$2,$3)
+fi
+])
+
 AC_DEFUN(RTEMS_CANONICALIZE_TOOLS,
 [AC_REQUIRE([RTEMS_PROG_CC])dnl
-if test "$rtems_cv_prog_gcc" = "yes" ; then
-  dnl We are using gcc, now ask it about its tools
-  dnl Necessary if gcc was configured to use the target's native tools
-  dnl or uses prefixes for gnutools (e.g. gas instead of as)
-  AR_FOR_TARGET=`$CC_FOR_TARGET --print-prog-name=ar`
-  AS_FOR_TARGET=`$CC_FOR_TARGET --print-prog-name=as`
-  LD_FOR_TARGET=`$CC_FOR_TARGET --print-prog-name=ld`
-  NM_FOR_TARGET=`$CC_FOR_TARGET --print-prog-name=nm`
-  RANLIB_FOR_TARGET=`$CC_FOR_TARGET --print-prog-name=ranlib`
-fi
 
-dnl check whether the tools exist
-dnl FIXME: What shall be done if they don't exist?
+dnl FIXME: What shall be done if these tools are not available?
+  RTEMS_PATH_TOOL(AR_FOR_TARGET,ar,no)
+  RTEMS_PATH_TOOL(AS_FOR_TARGET,as,no)
+  RTEMS_PATH_TOOL(LD_FOR_TARGET,ld,no)
+  RTEMS_PATH_TOOL(NM_FOR_TARGET,nm,no)
 
-dnl FIXME: This may fail if the compiler has not been recognized as gcc
-dnl       and uses tools with different names
-AC_PATH_PROG(AR_FOR_TARGET,"$program_prefix"ar,no)
-AC_PATH_PROG(AS_FOR_TARGET,"$program_prefix"as,no)
-AC_PATH_PROG(NM_FOR_TARGET,"$program_prefix"nm,no)
-AC_PATH_PROG(LD_FOR_TARGET,"$program_prefix"ld,no)
+dnl special treatment of ranlib
+  RTEMS_PATH_TOOL(RANLIB_FOR_TARGET,ranlib,no)
+  if test "$RANLIB_FOR_TARGET" = "no"; then
+    # ranlib wasn't found; check if ar -s is available
+    RTEMS_AR_FOR_TARGET_S
+    if test $rtems_cv_AR_FOR_TARGET_S = "yes" ; then
+      dnl override RANLIB_FOR_TARGET's cache
+      ac_cv_path_RANLIB_FOR_TARGET="$AR_FOR_TARGET -s"
+      RANLIB_FOR_TARGET=$ac_cv_path_RANLIB_FOR_TARGET
+    else
+      AC_MSG_ERROR([***]
+        [Can't figure out how to build a library index]
+	[Nether ranlib nor ar -s seem to be available] )
+    fi
+  fi
 
-dnl NOTE: This is doubtful, but should not disturb all current rtems'
-dnl 	  targets (remark: solaris fakes ranlib!!)
-AC_PATH_PROG(RANLIB_FOR_TARGET,"$program_prefix"ranlib,no)
-
-dnl NOTE: These may not be available, if not using gnutools
-AC_PATH_PROG(OBJCOPY_FOR_TARGET,"$program_prefix"objcopy,no)
-AC_PATH_PROG(SIZE_FOR_TARGET,"$program_prefix"size,no)
+dnl NOTE: These may not be available if not using gnutools
+  RTEMS_PATH_TOOL(OBJCOPY_FOR_TARGET,objcopy,no)
+  RTEMS_PATH_TOOL(SIZE_FOR_TARGET,size,no)
 ])
+
+AC_DEFUN(RTEMS_AR_FOR_TARGET_S,
+[
+AC_CACHE_CHECK(whether $AR_FOR_TARGET -s works,
+rtems_cv_AR_FOR_TARGET_S,
+[
+cat > conftest.$ac_ext <<EOF
+int foo( int b ) 
+{ return b; }
+EOF
+if AC_TRY_COMMAND($CC_FOR_TARGET -o conftest.o -c conftest.$ac_ext) \
+  && AC_TRY_COMMAND($AR_FOR_TARGET -sr conftest.a conftest.o) \
+  && test -s conftest.a ; \
+then
+  rtems_cv_AR_FOR_TARGET_S="yes"
+else
+  rtems_cv_AR_FOR_TARGET_S="no"
+fi
+  rm -f conftest*
+])
+])
+
+
+dnl Detect the Cygwin32 environment (unix under Win32)
+dnl
+dnl 98/06/16 David Fiddes   	(D.J.Fiddes@hw.ac.uk)
+dnl				Hacked from automake-1.3
+
+# Check to see if we're running under Cygwin32, without using
+# AC_CANONICAL_*.  If so, set output variable CYGWIN32 to "yes".
+# Otherwise set it to "no".
+
+dnl RTEMS_CYGWIN32()
+AC_DEFUN(RTEMS_CYGWIN32,
+[AC_CACHE_CHECK(for Cygwin32 environment, rtems_cv_cygwin32,
+[AC_TRY_COMPILE(,[return __CYGWIN32__;],
+rtems_cv_cygwin32=yes, rtems_cv_cygwin32=no)
+rm -f conftest*])
+CYGWIN32=
+test "$rtems_cv_cygwin32" = yes && CYGWIN32=yes])
+
+
+dnl Set the EXE extension
+dnl
+dnl 98/06/16 David Fiddes   	(D.J.Fiddes@hw.ac.uk)
+dnl				Hacked from automake-1.3
+
+# Check to see if we're running under Win32, without using
+# AC_CANONICAL_*.  If so, set output variable EXEEXT to ".exe".
+# Otherwise set it to "".
+
+dnl RTEMS_EXEEXT()
+dnl This knows we add .exe if we're building in the Cygwin32
+dnl environment. But if we're not, then it compiles a test program
+dnl to see if there is a suffix for executables.
+AC_DEFUN(RTEMS_EXEEXT,
+[AC_REQUIRE([RTEMS_CYGWIN32])
+AC_MSG_CHECKING([for executable suffix])
+AC_CACHE_VAL(rtems_cv_exeext,
+[if test "$CYGWIN32" = yes; then
+rtems_cv_exeext=.exe
+else
+cat > rtems_c_test.c << 'EOF'
+int main() {
+/* Nothing needed here */
+}
+EOF
+${CC-cc} -o rtems_c_test $CFLAGS $CPPFLAGS $LDFLAGS rtems_c_test.c $LIBS 1>&5
+rtems_cv_exeext=`echo rtems_c_test.* | grep -v rtems_c_test.c | sed -e s/rtems_c_test//`
+rm -f rtems_c_test*])
+test x"${rtems_cv_exeext}" = x && rtems_cv_exeext=no
+fi
+EXEEXT=""
+test x"${rtems_cv_exeext}" != xno && EXEEXT=${rtems_cv_exeext}
+AC_MSG_RESULT(${rtems_cv_exeext})
+AC_SUBST(EXEEXT)])
+
 
 dnl $Id$
